Refactor file IO Utils and address path security concerns
